+++ index.js
@@ -1,7 +1,9 @@
 exports.Agent = require('./lib/Agent');
 
-exports.Transport = require('./lib/transport/Transport');
-exports.LocalTransport = require('./lib/transport/LocalTransport');
-exports.PubNubTransport = require('./lib/transport/PubNubTransport');
-exports.AMQPTransport = require('./lib/transport/AMQPTransport');
-exports.DistribusTransport = require('./lib/transport/DistribusTransport');
+exports.transport = {
+  Transport: require('./lib/transport/Transport'),
+  LocalTransport: require('./lib/transport/LocalTransport'),
+  PubNubTransport: require('./lib/transport/PubNubTransport'),
+  AMQPTransport: require('./lib/transport/AMQPTransport'),
+  DistribusTransport: require('./lib/transport/DistribusTransport')
+};